Add React Router & lowdb
