+++ package.json
@@ -8,9 +8,11 @@
     "dotenv": "^8.2.0",
     "gh-pages": "^2.1.1",
     "grommet": "^2.14.0",
+    "lowdb": "^1.0.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-input-slider": "^5.1.4",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.2.0",
     "react-spring": "^8.0.27",
     "reactstrap": "^8.2.0",
@@ -26,7 +28,13 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "shared-config"
+    ],
+    "rules": {
+      "additional-rule": "warn"
+    }
   },
   "browserslist": {
     "production": [
@@ -41,7 +49,14 @@
     ]
   },
   "devDependencies": {
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.5.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "prettier": "2.0.5"
+    "eslint-plugin-react": "^7.20.4",
+    "prettier": "^2.0.5",
+    "pretty-quick": "^2.0.1"
   }
 }
