+++ src/App.js
@@ -1,11 +1,10 @@
-import React, { useRef } from 'react';
-import Gallery from './components/Gallery';
-import Forms from './components/Forms';
-import Footer from './layout/Footer';
-import Header from './layout/Header';
-import Search from './layout/Search';
-import { GalleryProvider } from './components/GalleryContext';
+import React, { Suspense, lazy } from 'react';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import { Grommet } from 'grommet';
+import { GalleryProvider } from './components/GalleryContext';
+
+const Home = lazy(() => import('./pages/home'));
+const GalleryPage = lazy(() => import('./pages/gallery'));
 
 require('dotenv').config();
 
@@ -13,32 +12,25 @@ const theme = {
   global: {},
 };
 
-const scrollToRef = (ref) =>
-  window.scrollTo({
-    top: ref.current.offsetTop,
-    left: 0,
-    behavior: 'smooth',
-  });
-function App(props) {
-  const myRef = useRef(null);
-  const executeScroll = () => scrollToRef(myRef);
+function App() {
   return (
     <div className="App container-fluid bg-dark">
-      <Grommet theme={theme}>
-        <Header />
-        <GalleryProvider>
-          <Search>
-            <Forms scroll={executeScroll} />
-          </Search>
-
-          <div className="row">
-            <div ref={myRef} id="gallery" className="col bg-dark">
-              <Gallery />
-            </div>
-          </div>
-        </GalleryProvider>
-        <Footer />
-      </Grommet>
+      <Router>
+        <Suspense fallback={<div>Loading...</div>}>
+          <Switch>
+            <Grommet theme={theme}>
+              <GalleryProvider>
+                <Route exact path="/">
+                  <Home />
+                </Route>
+                <Route exact path="/gallery">
+                  <GalleryPage />
+                </Route>
+              </GalleryProvider>
+            </Grommet>
+          </Switch>
+        </Suspense>
+      </Router>
     </div>
   );
 }
