+++ src/pages/gallery/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function index() {
+  return <div>wip</div>;
+}
